move profile functions to utils
